Replace standard contributions banner
