+++ src/modules.ts
@@ -41,16 +41,6 @@ export const contributionsBanner: ModuleInfo = getDefaultModuleInfo(
     'banners/contributions/ContributionsBanner',
 );
 
-export const contributionsBannerVariantA: ModuleInfo = getDefaultModuleInfo(
-    'contributions-banner-variant-a',
-    'banners/contributions/variantA/ContributionsBannerVariantA',
-);
-
-export const contributionsBannerVariantB: ModuleInfo = getDefaultModuleInfo(
-    'contributions-banner-variant-b',
-    'banners/contributions/variantB/ContributionsBannerVariantB',
-);
-
 export const digiSubs: ModuleInfo = getDefaultModuleInfo(
     'digital-subscriptions-banner',
     'banners/digitalSubscriptions/DigitalSubscriptionsBanner',
@@ -73,8 +63,6 @@ export const moduleInfos: ModuleInfo[] = [
     liveblogEpic,
     liveblogCardIconsEpic,
     contributionsBanner,
-    contributionsBannerVariantA,
-    contributionsBannerVariantB,
     digiSubs,
     guardianWeekly,
     puzzlesBanner,
